CI: security PHPCS, JavaScript syntax check and release workflow
- PHPCS with WordPress.Security on inc/, templates/ and entry points; every violation is an error
- Fix all findings: escaped output (_e -> esc_html_e, wp_die messages), wp_unslash on request data, absint on id lists, wp_safe_redirect; verified false positives annotated with the reason
- Request text no longer saved with WordPress magic-quote backslashes
- bin/check-js.js: node --check on assets/js and on inline scripts in PHP files
- CI workflow: php -l on PHP 7.4 and 8.3, PHPCS, JavaScript check, PHPUnit
- Release workflow: tag, header Version and DBEM_VERSION must match and the README must have the changelog entry; attaches a ZIP with the db-event-manager/ folder
- .gitattributes excludes development files from the ZIP

+++ templates/admin/survey.php
@@ -9,7 +9,7 @@ $events = get_posts(array(
     'order'          => 'DESC',
 ));
 
-$selected_event = absint($_GET['event_id'] ?? 0);
+$selected_event = absint($_GET['event_id'] ?? 0); // phpcs:ignore WordPress.Security.NonceVerification.Recommended -- parametro di navigazione in sola lettura
 $event_title = '';
 $responses = array();
 $survey_fields = array();
@@ -60,7 +60,7 @@ if ($selected_event) {
             <p>
                 <?php printf(
                     esc_html__('Risposte: %d | Presenti (checked-in): %d | Iscritti totali: %d', 'db-event-manager'),
-                    count($responses), $total_checked_in, $total_regs
+                    (int) count($responses), (int) $total_checked_in, (int) $total_regs
                 ); ?>
             </p>
             <button type="button" class="button button-primary" id="dbem-send-survey" data-event="<?php echo esc_attr($selected_event); ?>" data-target="checked_in">
